Prevent an infinite render loop in Canvas when the viewport is resized while very short.
Summary: Invoking the `setDefaultHeight` state function was causing a re-render, which would then call it again with the same value, repeating indefinitely. A quick check to prevent setting it to the value it already contains prevents this. This fix also includes two unrelated, no-op lint fixes.

Test Plan: Resize the window such that its height is under ~200px on any page that contains a `<Canvas/>`. The very first re-render caused by this resize triggers a render loop that crashes the app in production. With this fix, it should no longer do that.

Reviewers: nserrino, michelle, #engineering, zasgar

Reviewed By: #engineering, zasgar

JIRA Issues: PC-578

Differential Revision: https://phab.corp.pixielabs.ai/D6392

GitOrigin-RevId: 0bfbc9b9827d323e37546edeeca446d392cc020b