Fix entity search breadcrumbs
Summary:
recent breadcrumb refactors broke the entity search (pod/service) breadcrumbs. when typing in the input box for these arguments, we need to make a call to the backend to get the list of suggestions.
the breadcrumbs component only calls getListItems when the function changes, but this needs to be called everytime the user types in the input box.

Test Plan:
ran in webpack and manually verified that we could search in all of the different types of breadcrumbs:
- cluster
- script
- entity breadcrumbs (pods/svcs)
- non-entity breadcrumbs (start_time)

Reviewers: nick, nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6401

GitOrigin-RevId: 62e5703199018ce1ee3f881aaa99618d6f0d5be2