PC-567: Divide udf w/ ints (1/2, etc) should output a float to match Python
Summary: This relates to PC-567 because we want throughput rates that might be less than 1 unit per second. We previously needed to multiply by 1.0 to get this behavior, but it messes up the semantic typing to do that. While looking into this, we noticed that since 1/2 in python is 0.5, we should have the same in pixie, which conveniently kills two birds with one stone.

Test Plan: edited unit test

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-567

Differential Revision: https://phab.corp.pixielabs.ai/D6408

GitOrigin-RevId: 6ad0dea0635051716c0d915f01f24ba40f3e8376