[PC-601] Fix autocomplete situation where input would be ignored on the second usage
Summary:
When we load a page, the autocomplete starts with tab stops in a non-empty state that somehow come from graphql. When you submit the autocomplete, that state would be wiped away and reset with an empty array. That means the next invocation of the autocomplete, you have to do a graphql request after typing characters, which will continue to happen until you stop typing.

This interaction was really annoying for a while and meant that if you wanted to write a script quickly, you'd find yourself 4-5 characters down the string you were trying to type and only then would the characters start registering.

The better solution is to accept inputs before graphql calls, but tbh it might be too messy for me to address.

Test Plan: Tested on local dev ui and fixes the problem

Reviewers: michelle, nick, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-601

Differential Revision: https://phab.corp.pixielabs.ai/D6407

GitOrigin-RevId: 6e47cfdff785f9a2c51c6d5423e146c2e0e2e029