Fix bug where autocomplete backend reports some valid command as invalid
Summary:
the autocomplete can incorrectly determine a command as invalid if it hits this case:
consider the command: "script:px/pod pod:pl/pod_name"
elastic may give suggestions for the pod as:
pl/pod_names
pl/pod_name
pl/pod_name is a valid entity, but because it's not the first, elastic will consider the command as invalid.
this makes the interaction weird, because when you press enter the command won't execute.
instead, the command should be valid if it matches any of the suggestions.

Test Plan: n/a, this is a bit difficult to add in our elastic suggester test since we don't actually want a test to depend on how elastic orders things, as this may change.

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6409

GitOrigin-RevId: 738129ee6924343293c4b4c7bbabb5777da88550