If pxl contains mutation, show different message when no tables exist
Summary:
it is confusing when we open a dynamic tracing script, because we don't automatically execute it and it shows ""x" not found"
instead we want it to say "Run to generate results".
Note that if, for some reason, the dynamic tracing script actually did not generate an expected table, it will still say "Run to generate results". we will need some more statepassing to know whether or not a script has actually been executed or not, but that is probably too involved for tonight

Test Plan:
ran in webpack
screenshot:
{F96397}

Reviewers: nserrino, zasgar, philkuz, #engineering

Reviewed By: nserrino, philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6421

GitOrigin-RevId: fceb218a93c390de2893176e0b907a3da4711596