Update redeploy_etcd for etcd-operator to also bounce etcd-operator/pl-etcd pods
Summary:
we're seeing some issues with the customer cluster failing to connect to etcd. we have tried redeploy etcd, but that doesn't appear to make a difference.
currently, the redeploy_etcd process for clusters running with etcd-operator, as opposed to etcd statefulset, deletes the pl-etcd (etcd crd). this clears out the etcd and starts a clean new instance. this is what we usually need in cases where we have a bad set of data in etcd.
however, this leaves the etcd-operator and its pl-etcd pods in the pl namespace. the operator picks up the pl-etcd change and things usually proceed as normal.

the situation occurring in customer may indicate that there may be some issue with the etcd-operator or pl-etcd pods. i updated the job to restart those as well, in case that is where the issue lies.
note: there is no guarantee that this will fix the customer issue, but it is worth a shot, and doesnt hurt to have in the case where we do need to restart etcd-operator/pl-etcd.

Test Plan: started an rc, launched a cluster running etcd_operator, verify that updating to the rc with --redeploy_etcd flag does delete the operator, any running pl-etcd pods, then relaunches them. verified that metadata is able to reconnect and Pixie starts up as normal.

Reviewers: zasgar, jamesbartlett, #engineering, nserrino

Reviewed By: #engineering, nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D6423

GitOrigin-RevId: afda103c648e664d47ebeaa5fbf4246d2ea72013