Do not assume <input/> references are available when trying to autofocus.
Summary: This was triggering errors in a state update, which in turn causes a component to leak. This is usually harmless, but a memory leak is a memory leak and log spam is annoying.

Test Plan: Check the console when closing the breadcrumbs after editing them early into the page load. There are other triggers, but that one's easy to reproduce.

Reviewers: nserrino, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6424

GitOrigin-RevId: 81f656379c241d24af681f91b5811b148cc42055