Cleanup UI healthcheck code
Summary:
we investigated a bug where the UI would get stuck in the "Connecting to cluster" modal.
it turns out there is race condition when the useEffect which handles our healthcheck code is called in quick succession. this can happen when the clusterID changes quickly.
this happens in the repro steps described in the test plan, but could possibly occur from other interactions as well.

in the useEffect, we set SubscriptionRef in a promise. if the clusterID changes before the promise is done, the useEffect cleanup will not properly cleanup the subscription and it runs forever. this leads to a state where we have two running subscriptions for two different clusters that keep overwriting each other.
cleaned up the code so that each useEffect call is responsible for its own subscription/retry, since these don't need to be in the react state.

also noticed a bug that occurs when switching between clusters normally... we never actually cancel the healthcheck grpc stream. so if you switch between 5 different clusters, you would end up with 5 healthcheck streams running at the same time. updated the healthcheck so that it is cancel when unsubscribed.

Test Plan:
normally switch between clusters, verify the GRPC-web stream looks as expected and that we can run scripts on the current cluster.
tried to repro the bug:
- have one tab open to cluster A
- click a link that opens a tab to cluster B
- verify that we end up on cluster B and can run scripts as normal

Reviewers: nserrino, nick, #engineering

Reviewed By: nick, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6426

GitOrigin-RevId: 9c20fdda8ac3ccad2c297b8a2d69883e9ea2a177