Show vizier Update job pod in Control Plane Pods
Summary:
some customers take a really long time to update, maybe due to PodScheduling or other constraints.
sometimes when we look at the control plane pods and there are some pods missing, we're unsure if its still in the update process or if something has actually broken.
this diff updates the update job label so that it is tracked in the control plane pods.

Test Plan: built an rc, updated my cluster and checked that i see the update job in the control plane pods

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6427

GitOrigin-RevId: cfb573833ed56de35e9ec533c4fa9ba9227a1c2a