PC-581 Fix breadcrumbs overflow model in browser
Summary:
currently when the breadcrumbs overflow, we expect the user to be able to scroll horizontally to view all breadcrumbs.
this is a confusing interaction in desktop browsers. instead, in a desktop browser, we want to let the breadcrumbs stack so that they are all viewable without any scrolling.
we should keep the previous interaction for a mobile-sized screen.

Test Plan:
Stacked breadcrumbs in browser:
{F96639}
Regular breadcrumbs that fit:
{F96640}
Breadcrumbs in a mobile-sized screen:
{F96641}

Reviewers: zasgar, nick, nserrino, #engineering

Reviewed By: nick, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6447

GitOrigin-RevId: be56675330427a1c06648e4d2a4bb16700bae592