PC-567: Refactor stream_adapter in CLI code to have a separate component for data formatting.
Summary:
Previously the data formatting was integrated in stream_adapter.go. This diff adds a formatter class that is a bit easier to test and extend.
This diff adds support for the new semantic types that we have in the system, but some of them won't work until all of our pxl scripts are updated to use them.
It adapts the previous stream_adapter code to be more similar to what we do in the UI.

Test Plan: added a unit test

Reviewers: zasgar, jamesbartlett, philkuz, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-567

Differential Revision: https://phab.corp.pixielabs.ai/D6440

GitOrigin-RevId: a833184fbfe9c0f3d11de4e2f69145ccb61f579e