[PC-549] Fix errors not showing up when the VisJSON is formatted incorrectly
Summary:
We have this issue where  VisJSON incorrectly formatted looks like it works. In reality the parser would error out, the error would be caught and not handled, and we would use the old vis value.

Now replaced all parseVis w/ parseVisOrShowError which shows an error if the parseVis fails.

Test Plan: tested on yarn dev, not sure how to properly test this.

Reviewers: nick, michelle, nserrino, #engineering

Reviewed By: nserrino, #engineering

JIRA Issues: PC-549

Differential Revision: https://phab.corp.pixielabs.ai/D6449

GitOrigin-RevId: 5eb2374fd49906cd9673c5174d544264b6198245