PC-567: ST_DURATION_NS_QUANTILES should render duration-formatted box whisker quantiles
Summary: We already have box whiskers for ST_QUANTILES, which just has the quantiles stored as plain floats. this diff adds support for quantiles where the values are known to be durations in nanoseconds. It also provides contextual highlighting when the column has 'latency' in the name.

Test Plan: added/edited unit.

Reviewers: zasgar, jamesbartlett, nick, philkuz, #engineering

Reviewed By: jamesbartlett, #engineering

JIRA Issues: PC-567

Differential Revision: https://phab.corp.pixielabs.ai/D6464

GitOrigin-RevId: 21d3b700793108496265298988718b3253ae60eb