Unify the format-data formatting and vega-spec formatting to simplify future SemType additions.
Summary: Needed to add semtypes to the vega charts for thorughput_per_ns and bytes_per_ns so I unified the logic to only need to update a formatter in one place. Also had to add logic to avoid min_domain_scaling if a formatter is used for the axis.

Test Plan: Tested on yarn dev, see #ui channle

Reviewers: jamesbartlett, zasgar, nick, #engineering

Reviewed By: jamesbartlett, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6465

GitOrigin-RevId: ab3f3cd5003cf590927b59531f003e7993f6c19f