Use pretty printed tables in the data drawer (formatted data).
Summary:
This should probably be the default behavior with semantic types.

{F96905}

Test Plan: Open the data drawer for a script that produces tables with pretty-printed columns, for instance `px/cluster` in the `gke:prod` cluster. It should pretty print the data for each table the same way as they do outside of the drawer.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-626

Differential Revision: https://phab.corp.pixielabs.ai/D6472

GitOrigin-RevId: 855ced91bece6f9e04ada9cce20b77ba7e73d644