Deduplicate items in breadcrumb dropdowns, to prevent bizarro states with multiple (duplicate) items highlighted.
Summary: Per the inline TODO, this is temporary code. The API should not be providing duplicates in the first place, but it was faster to implement a dedup in the frontend while that is being worked on. Launch moooooonth!

Test Plan: Try the `px/services` script's `service` dropdown.

Reviewers: nserrino, michelle, #engineering, philkuz

Reviewed By: #engineering, philkuz

Subscribers: philkuz

JIRA Issues: PC-630

Differential Revision: https://phab.corp.pixielabs.ai/D6477

GitOrigin-RevId: 58b50fa7f1f11ca4490794da7cf7a799e172438a