Refactor CLI LoadDefaultCredentials to show better error message when user should be logged in
Summary:
Most of our CLI operations require the user to be logged in. however, a lot of the times we don't print a very friendly error message to the user.
Created a MustLoadDefaultCredentials function so that we print a better message for the user, and exit out of the CLI. We keep LoadDefaultCredentials around for the cases where we don't actually need the user to be logged in (when we check for whether or not the user's CLI can be updated or not).

Test Plan: ran the CLI unauthenticated and authenticated

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6485

GitOrigin-RevId: ca25cb904725cbdf8afb6d5b2833a3bd29661647