Turn "Run" button into "Stop" button when a script is running long.
Summary:
It waits one second before doing this to avoid flickering on scripts that run quickly.

{F97120}

Test Plan: Run a fast script, a slow one, a few that fail reliably, etc. For one that usually takes a long time, try `etcd_get_trace`; `http_data` in production sits close to the threshold for some clusters and is a good example for both cases.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-635

Differential Revision: https://phab.corp.pixielabs.ai/D6493

GitOrigin-RevId: bbca0b6da6270b8f72bee45105f7313db53494a5