Fix analytics for login
Summary:
We were not properly tracking user logins/identifies, since we redirect too quickly after the analytic functions are called.
instead, we need to actually wait for the call to segment to be complete before redirecting. unfortunately, the best way to do this is to wrap their function in a promise and resolve the promise in the callback: https://github.com/segmentio/analytics-node/issues/93

now that we are waiting for the segment call to be made, we actually stay on the "Logging in..." page for a few seconds. (the page in the screenshot below)

Test Plan:
ran in webpack while checking segment debugger
{F97130}

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6494

GitOrigin-RevId: 7b2151a6a4352e4b57960f456814e00ab083809d