SocketTracer: Use byte positions instead of sequence numbers
Summary:
Use byte position instead of sequence numbers for BPF events.

For now, this is just a different way to encode events, and may seem like a lateral move rather than an improvement.

This will, however, open the door to putting the data directly into a buffer instead of an std::map and then copying the data to process it.

Test Plan: Existing tests.

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

JIRA Issues: PP-935

Differential Revision: https://phab.corp.pixielabs.ai/D6491

GitOrigin-RevId: dc7dc145ecd6f7fdbeae4fc76847581270d46d69