Fix whitescreen from request-graph namespace parsing
Summary:
there are two issues here:
1. a race condition when the script is changed
when a script is selected from the breadcrumbs, we call setScript to set the pxl/vis/args etc. This immediately triggers a rerender of the canvas, which actually tries rendering the request graph with the new display with old data.
however, we noticed that the script change using cmd-k doesn't try to render the request graph with old data.
the script switch code looks the same. however, for cmd-k, after setScript is called, it also calls execute() before the rerender of the canvas is triggered. execute sets loading to true on the canvas, so it doesn't try to render the request graph with old data.
to handle this, i updated the ordering of the script/execute, since the loading state should always be set before the canvas rerenders and execute has no dependency on the current script state.
this fix is not actually necessary because of the fix for (2) below. it is also not the greatest fix, because what we should really do is cleanup our UI and all of the useEffect chains which are causing unnecessary renders

2. the request graph should be able to handle an invalid data and should never whitescreen in the first place.

Test Plan: switched between scripts in both the breadcrumbs and cmd-k

Reviewers: zasgar, nserrino, #engineering, philkuz

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6445

GitOrigin-RevId: 812e4773981c5ffb0951e8163c3a66f05f4ce0e6