Don't replicate a singleton unload handler.
Summary: It only needs to run once and it acts on a stateful variable, so there's no point repeating the listener (nor removing an old one to do so).

Test Plan: Set a debugger on live.tsx:163. In prod, it triggers many times if you've run a few scripts or switched clusters several times in a run. This fix makes it run once.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6495

GitOrigin-RevId: 1e8aeda65823338362eb1e0a0263cd38eca31a8b