Override width for duration columns
Summary:
This is imperfect, but it gets the job done. See screenshots.

On the provided script in PC-628:
{F97174}

On a smaller screen:
{F97176}

On the default `px/http` script:
{F97175}

Test Plan: See PC-628 for example script. Also test with the default `px/http_data`, and using the mobile dimensions emulation in your browser. In all cases, the default width of `ST_DURATION_NS` columns should now be more reasonable.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-628

Differential Revision: https://phab.corp.pixielabs.ai/D6498

GitOrigin-RevId: 5ee77b244c398def1020268431e8aeb3a41cd3e7