Send K8s Events in Cloudconn heartbeats
Summary:
We've been seeing the customer clusters struggling to start up without any sentry errors or logs in elastic.
by showing the k8s events in our control plane UI, similar to a `kubectl describe pod`, we should be able to narrow down whether there is an issue with starting the pod in K8s itself.
this diff just adds the k8s events in the heartbeat.
after this, there also needs to be a diff on the cloud-side for making this data available over gql.
once that is done, we can add the information in the UI

Test Plan: updated 0.5.2 -> an rc, to make sure the cloudconn role gets updated successfully by the update job. check that the rc can get events from k8s API

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6502

GitOrigin-RevId: 47a400699e8a3211ad74ae6d4c7016fa0d12b217