Move announcekit from behind feature flag
Summary:
before, we put announcekit behind a featureflag because we weren't ready to show it to everyone yet.
now, we want to actually start sending announcements, and it seems unlikely that we'll remove announcekit in the future, so it doesn't need to be behind a featureflag anymore.

Test Plan: ran webpack and check that announcekit does now show up for everyone

Reviewers: zasgar, nserrino, nick, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6508

GitOrigin-RevId: d025ca99d5dbf2a7277e46f4545658ac12005fb5