Make k8s events available on PodStatusResolver
Summary:
in a previous diff, we updated the PodStatus so that it also includes k8s events. this should give us more information to why some customer clusters are failing to start up without any sentry errors/elastic logs.
this diff makes the k8s events available on graphql resolver so that we can fetch it in the UI

Test Plan: unit tests

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6511

GitOrigin-RevId: 63a1cc5a6c6e396f2a8e365bc65e5254df1f54d3