Purge localStorage and sessionStorage on user logout.
Summary: This needs to be done for several reasons, but the impetus was seeing Bolt clusters when a support account was looking at customer clusters. The issue was that Apollo's cache was being consulted while a graphql query was stalling due to a partial node outage; the Apollo cache shouldn't be persisting between logins in the first place.

Test Plan: Look at the clusters page when your devtools are set to use an extremely high-latency connection. Heartbeats won't update every 2.5sec like they should, because latency is to high to do so. On a fresh login, you should no longer see stale data while waiting out the latency period.

Reviewers: nserrino, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6515

GitOrigin-RevId: 58901205697e8ecb42b6ed64a85bc0cea48631b4