Add virtual destructors to other exec nodes.
Summary: In D6531, @zasgar addressed a memory leak that we were seeing in Kelvin and PEM nodes. It was due to the lack of a virtual destructor in the agg node. In this diff, I add virtual destructors that were missing to our other exec nodes.

Test Plan: n/a

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D6532

GitOrigin-RevId: d21e5b655dfeaab58369f662df460022c1781f07