Update Admin UI pod count to show count of running PEMs
Summary:
we updated PEMs so that they can be deployed on tainted nodes. this started showing us weird results in the admin UI, where it says the # of instrumented nodes is greater than the number of nodes actually in the cluster.
looking at the count of agents in px/agent_status actually shows a more correct number of instrumented nodes. i wasn't able to repro it on our own clusters, but my guess is that there are some PEMs stuck terminating, or stuck trying to be scheduled, but we are counting these anyway

Test Plan: ran in skaffold, checked the PEM count

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6551

GitOrigin-RevId: 93ad0d1cb315bc248d53af954649cec4829e780c