Static Tour Overlay (WIP)
Summary:
Currently only opens from a new button in the sidebar. A future change can make it appear once automatically per user. The button does not appear on mobile (viewport widths under 960px) as the tour does not fit there and the layout is different anyway.

When it's open:
{F103838}

Test Plan: Open the profile/admin menu in the bottom left in the sidebar. It should contain a `Tour` item that should open the tour.

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: htroisi

JIRA Issues: PC-426

Differential Revision: https://phab.corp.pixielabs.ai/D6557

GitOrigin-RevId: a8a8670fc811314ff2fc82478d0a2a00295ca2ee