Simplify wrapping layers of login/logout/callback pages with Flexbox.
Summary: Part one of two. Second part (at a later date) will be doing the same to the login/logout/callback contents themselves.

Test Plan: Load up the sign-in/sign-up pages in a browser on a desktop and/or phone. Resize the window to your heart's content. Note that everything stays on-screen, centered when appropriate, and scrollable when necessary, no matter the screen dimensions.

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-645

Differential Revision: https://phab.corp.pixielabs.ai/D6558

GitOrigin-RevId: 36d2a67d31c66a430297eb245e225ef39fca6e83