[PP-2252] Support executing UDF in the compiler for constant values
Summary:
Reworked the previous version of this Diff so that we can load UDFs in the compiler. Currently only load the math ops but we can extend it further.

In the visitor, when you find a function on constant arguments, finds a matching UDF, applies the UDF, and return the constant value.

Test Plan: Wrote tests and had to update some AST_visitor tests because we change a bit of how things work.

Reviewers: nserrino, jamesbartlett, #engineering, zasgar

Reviewed By: jamesbartlett, #engineering, zasgar

JIRA Issues: PP-2252

Differential Revision: https://phab.corp.pixielabs.ai/D6538

GitOrigin-RevId: 4896532637ac07abcdd30a658bbb5ae68e0fd507