Add DebugLogServer in Vizier
Summary: we want Vizier to be able to respond to DebugLogRequests. this arrives as a passthrough request, which is picked up by querybroker who handles passthrough requests. querybroker makes a call to the cloudconnector who makes the call to get the k8s log for the given pod.

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6585

GitOrigin-RevId: 41bd612e0744279e1d171ab45a1274816ba23327