Use a smaller CIDR range for POD addresses
Summary:
The default pod range is `/14` and I believe that is what was preventing me
from being able to create a gke cluster. Set it to a much smaller `/21` which should
be sufficient for dev purposes.
Also explicitly specify the services range, set to the default value of `/20`

Test Plan: `./scrips/create_gke_cluster.sh` now works.

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6603

GitOrigin-RevId: 7eadb93430422a446ef6db475e75d4843efd314f