Choose kubeconfig as default cluster, then search for first healthy vizier cluster if that fails
Summary: Figured we should use the one that's in the kubeconfig, most people in the cli  probably have kubectl or they will specify the cluster ID anyways. This simplifies it.

Test Plan: Tested when the cluster was healthy (queried the kubeconf cluster). When unhealthy, searched for a healthy cluster (aka fall back on old behavior).

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6601

GitOrigin-RevId: c2633d5a0525c1cd3acfbb1f12694e50226a0176