Fix hanging health checks: cancel query after healthCheckInterval
Summary:
We noticed that health checks were starting to get stuck in customer.
Upon closer investigation, this would always occur after the querybroker logged: `rpc error: code = Unauthenticated desc = invalid auth token: Token used before issued`
To try to repo in my own environment, I updated exec stats to initiate the TransferResultChunk stream with a jwt with an IAT in the future. this was able to reproduce the issue.
after adding in tons of logs, it looks like our healthcheck is getting stuck in qb's `StreamResults`, which keeps waiting forever until the query is done or has been cancelled. however, it never receives anything and waits forever because once carnot fails to send the ExecStats, it doesn't try again.
we already have a healthCheckInterval which should unblock the healthcheck after it waits too long. we just need to make sure this also cancels the query, in case it has gotten stuck.

as for why customer hit the "Token used before issued" problem, I'm guessing there is possibly some small clock skew between the kelvin and qb on their cluster.

Test Plan:
had carnot try to send exec stats with a token with a IAT in the future. verified that this causes the healthchecks to hang while waiting for StreamResults to finish.
Added in my change, and verify that health checks no longer hang.

Reviewers: zasgar, nserrino, #engineering, vihang

Reviewed By: vihang

Subscribers: vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6627

GitOrigin-RevId: 9fc9ae9664491590c568a9adeb5422d827c7fe0c