Propagate PodConditions to our metadata tracker and use them
Summary:
The PodConditions include all condition types and states for each condition,
so this goes from holding on to just a repeated set of conditions, to
conditions and their statuses, and propagating the same down to the metadata state.

Also add a UDF to get pod ready state. This UDF just returns a bool that checks for
the PodCondition Ready and returns false if the state is either unknown or false.
We can consider beefing up this UDF to instead return all possible condition states
(true/false/unknown) and letting the user query either of the known states.

Test Plan: Added unit tests wherever applicable. Would love help integration testing it on my dev machine.

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: nserrino, #engineering

JIRA Issues: PP-2079

Differential Revision: https://phab.corp.pixielabs.ai/D6628

GitOrigin-RevId: 1c615bf521b668a19278224df78c16ea6db6f19a