Fix bug where vizier auto-updater can update multiple times
Summary:
we've seen cases where the vizier auto-updater triggers another update after a vizier has already finished updating.
here's what was happening:
- a vizier sends a heartbeat. vzmgr checks the version of that vizier. if an older version is running, the autoupdater queues up the vizier for an update.
- autoupdater sends out the update to that vizier. vzmgr sets the vizier's status as UPDATING in the db.
- vizier receives the update, starts updating.
- while updating, vizier is still sending heartbeats. at this point, we werent yet tracking the update job, and would just send "HEALTHY" to vzmgr
- vzmgr updates the status to "HEALTHY" in the db in the next heartbeat.
- vzmgr sees in the heartbeat that the vizier is healthy, and still at an older version
- autoupdater queues the vizier for another update
This mainly happened because we updated the update job so that it deletes the cloudconn as the very last thing it does in the updater job. before, it used to be the very first thing the updater job did. we made this change so that theres always a live cloudconn instance running in case the update job fails.
now, the vizier is correctly in "UPDATING" status whenever the update job is actually running.

Test Plan: created rc, updated to that rc, then updated that rc

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6635

GitOrigin-RevId: 73f0b21829f19a27f217440b8d61a0a3aa8ed2f5