Better algorithm for table column resizing and limits
Summary: Resizing any column should now be substantially more intelligent. Columns will do their best not to change sizes outside of the one being dragged, but they will yield if they need to. The table as a whole shrinks and grows to help with this.

Test Plan: Open a script with a good number of columns like `px/http_data` and play around with resizing columns. Do the same for a table with fewer columns, like `px/agent_status`. See implementation and unit tests for full details on how it should behave. KNOWN ISSUE: For the latter, the first movement jumps all of the column widths to something more uniform. That will be investigated in a future change.

Reviewers: nserrino, michelle, vihang, #engineering

Reviewed By: vihang, #engineering

JIRA Issues: PC-655

Differential Revision: https://phab.corp.pixielabs.ai/D6747

GitOrigin-RevId: 6f8d4828a7b251a35e4ef819ad106aeb7dcc0694