Add bazel build rule for building cloud tar bundle
Summary:
for our due diligence, we want a tar of all of our cloud images. we already have this command for our vizier images, but not for our cloud images since they are deployed through skaffold.
We may not ever need this command again, but might as well check it in so we don't have to redo the work if we ever need this again

Test Plan: ran `bazel build //k8s/cloud:cloud_image_bundle.tar`

Reviewers: zasgar, vihang, #engineering

Reviewed By: vihang, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6663

GitOrigin-RevId: 3254e6e820b1d2af71a135733ab416e31ee6797f