Cleanup licenses build rules
Summary:
This does some refactoring in an effort to clean dupe and fix the Jenkins build.

If we are running with stamping enabled, treat missing licenses as fatal
failres, otherwise just write to the missing licenses files and move along.

This should prevent the build from failing if we cannot SOPS decrypt the github api token
(unless we are stamping the build for release, in which case, the credentials to SOPS
decrypt need to be threaded through).

Test Plan:
Run the build locally with and without the glcoud application-default token.
Also hopefully this fixes Jenkins builds off main.

Reviewers: michelle, philkuz, #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D6677

GitOrigin-RevId: e8e654ee4b6a52fd2bd83d9860d3d54e88481b19