Add operator pruning for ASID filters
Summary:
the Heap UDTFs raised the issue that we don't support ASID filters at the moment. This makes it so that we can filter if you use px.asid(). Not completely sure if I want to allow sweeping filter on asid. I'm worried about a situation like data w/ an asid column is stored on a Kelvin but was originally collected on a PEM. Filtering on an asid column in this case would mean we would prune hte Kelvin part which is wrong. I tried to make this as explicit as possible by only allowing
`px.asid() == 123` as the valid asid filter.

After a suggestion from Natalie, I've generalized the approach to create a Matcher factory that matches specific sub expressions. Users can add a matcher function to the metadata factory, the factory extracts the MatchExpr() static function. Whenever a match is made, factory creates the matcher operation and then Gets the Agents that should be pruned for the expression.

As before, we continue to support logical operators `and`/`or` on top of these fine-grained expressions.

Test Plan: Wrote a test for explicit case which this supports and against the implict case which this does not support.

Reviewers: nserrino, #engineering, vihang

Reviewed By: #engineering, vihang

Differential Revision: https://phab.corp.pixielabs.ai/D6647

GitOrigin-RevId: 8396edce0ff5f41c091503bb546ae1e0dc59bc99